# docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  controllers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controllers.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  controllers.ResetPasswordRequest:
    properties:
      email:
        type: string
      new_password:
        maxLength: 50
        minLength: 6
        type: string
      otp:
        type: string
    required:
    - email
    - new_password
    - otp
    type: object
  controllers.VerifyOTPRequest:
    properties:
      email:
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
...
  domain.User:
    properties:
      _id:
        description: |-
          The unique identifier of the user
          example: 507f1f77bcf86cd799439011
        type: string
      bio:
        description: |-
          User biography or description
          example: Developer and blogger
        type: string
      contact_info:
        description: |-
          Contact information (phone, etc.)
          example: +1234567890
        type: string
      email:
        description: |-
          Email address of the user
          example: johndoe@example.com
        type: string
      full_name:
        description: |-
          Full name of the user
          example: John Doe
        type: string
      google_id:
        description: |-
          Google account ID for OAuth login
          example: 1234567890abcdefg
        type: string
      is_verified:
        description: |-
          Whether the user has verified their email
          example: true
        type: boolean
      otp_code:
        description: OTP code for verification
        type: string
      password:
        description: |-
          Password (hashed) for authentication
          required: true
          min length: 6
          max length: 50
          example: $2a$12$eXampleHashedPassword...
        maxLength: 50
        minLength: 6
        type: string
      picture:
        description: |-
          URL to the user's profile picture
          example: https://example.com/images/johndoe.jpg
        type: string
      profile_picture:
        description: |-
          Profile picture URL
          example: https://example.com/images/profilepic.jpg
        type: string
      refresh_token:
        description: Refresh token for session management
        type: string
      reset_otp:
        description: Reset OTP code for password reset
        type: string
      role:
        description: |-
          Role of the user (e.g., user, admin)
          example: user
        type: string
      username:
        description: |-
          Username of the user
          required: true
          min length: 3
          max length: 50
          example: johndoe123
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the Blog API documentation.
  title: Blog API
  version: "1.0"
paths:
...
  /users/demote/{id}:
    post:
      description: Lowers a user's role to regular user, requires admin privileges
      parameters:
      - description: Target user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User demoted to regular user
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Demote admin to regular user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticates user credentials and returns access and refresh tokens
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Access and refresh tokens
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user and return JWT tokens
      tags:
      - users
  /users/logout:
    post:
      description: Invalidates user's refresh token to logout
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Logout failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Logout user
      tags:
      - users
  /users/profile:
    put:
      consumes:
      - application/json
      description: Modifies the user's profile information
      parameters:
      - description: Updated user profile data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user profile
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update profile
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update user profile
      tags:
      - users
  /users/promote/{id}:
    post:
      description: Elevates a user's role to admin, requires admin privileges
      parameters:
      - description: Target user ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User promoted to admin
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Promote user to admin
      tags:
      - users
  /users/refresh-token:
    post:
      consumes:
      - application/json
      description: Generates a new access token using a valid refresh token
      parameters:
      - description: Refresh token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/controllers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New access token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - invalid token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Registers a new user and sends an OTP to verify email
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent to your email. Please verify.
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user and send OTP for email verification
      tags:
      - users
  /users/reset-password:
    post:
      consumes:
      - application/json
      description: Verifies OTP and updates user's password
      parameters:
      - description: Reset password details
        in: body
        name: resetRequest
        required: true
        schema:
          $ref: '#/definitions/controllers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input or reset failed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset user password
      tags:
      - users
  /users/send-reset-otp:
    post:
      consumes:
      - application/json
      description: Sends a one-time password to user's email for password reset
      parameters:
      - description: Email to send OTP
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/controllers.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Reset OTP sent to your email address
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to send reset OTP
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send password reset OTP
      tags:
      - users
  /users/verify-otp:
    post:
      consumes:
      - application/json
      description: Verifies a user's email by OTP
      parameters:
      - description: Email and OTP
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controllers.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Account verified successfully!
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Verification failed or bad input
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify email using OTP
      tags:
      - users
swagger: "2.0"